Create a driver program to run problem modules.
This driver program will prompt users to enter problem. With the input will then
dynamically load the module and then run the required functions
`part_one_answer` and `part_two_answer`. The returned values will then be
displayed in the terminal.

The reason why this was implemented was to provide a better user interface
for users to run and view my problem answers, without having to run the module
directly.