Reformat problem 16 code solution.
Cleaned up the problem 16 code solution to abide by black and flake8 pre-commit
hooks. Also, added the part_one_answer and part_two_answer public functions to
be able to run by the main program.

However, the part_one_answer function does not return the answer because of
the nature of my solution. A fix shall be implemented in the future.